Compilers suck.

+++ Malloc1/fb4.c
@@ -18,7 +18,7 @@ int main()
      Print out the integers before the return values. */
 
   i = 1000;
-  printf("sbrk(0) = 0x%x\n", sbrk(0));
+  printf("sbrk(0) = 0x%x\n", (unsigned int) sbrk(0));
   for (sz = 4; sz < 32; sz += 4) {
     buf = (int *) malloc(sz);
     buf[0] = i;
@@ -35,7 +35,7 @@ int main()
   i++;
   printf("Allocated %d bytes.  buf = 0x%x, buf[-1] = %d, buf[-2] = %d, buf[0] = %d\n", 
             sz, (unsigned int) buf, buf[-1], buf[-2], buf[0]);
-  printf("sbrk(0) = 0x%x\n", sbrk(0));
+  printf("sbrk(0) = 0x%x\n", (unsigned int) sbrk(0));
 
   return 0;
 }
